+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Feat: match longest definitions first. See https://github.com/DiscreteTom/bimark/issues/3.
 - Feat: ignore files in `.gitignore`.
+- **Breaking Change**: remove `triggerCharacters`.
 
 ## v0.1.0